4i_analysis/4i_processing/dev-requirements.txt
numpy
matplotlib
scikit-image
pandas
tqdm
scipy
pyyaml
setuptools
cellpose
igraph
scanpy
anndata
itk-elastix
louvain
PyWavelets
tifffile
smo